# packages/api/openapi.json
{"openapi":"3.1.0","paths":{"/v1/auth/register":{"post":{"operationId":"registerV1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterBody"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}}},"tags":["/auth"]}},"/v1/auth/login":{"post":{"operationId":"loginV1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginBody"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}},"tags":["/auth"]}},"/v1/projects":{"post":{"operationId":"createProjectV1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectBody"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectResponse"}}}}},"security":[{"bearer":[]}],"tags":["/projects"]},"get":{"operationId":"getAllProjectsV1","parameters":[],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAllProjectsResponse"}}}}},"security":[{"bearer":[]}],"tags":["/projects"]}},"/v1/projects/{id}":{"get":{"operationId":"getProjectByIdV1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetProjectByIdResponse"}}}}},"security":[{"bearer":[]}],"tags":["/projects"]},"put":{"operationId":"updateProjectV1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectBody"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectResponse"}}}}},"security":[{"bearer":[]}],"tags":["/projects"]},"delete":{"operationId":"deleteProjectV1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteProjectResponse"}}}}},"security":[{"bearer":[]}],"tags":["/projects"]}},"/v1/tasks":{"post":{"operationId":"createTaskV1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskBody"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskResponse"}}}}},"security":[{"bearer":[]}],"tags":["/tasks"]},"get":{"operationId":"getAllTasksV1","parameters":[{"name":"filter","required":false,"in":"query","schema":{"examples":[{"status":"todo","priority":"high","projectId":1}],"type":"object","properties":{"status":{"anyOf":[{"const":"todo","type":"string"},{"const":"in_progress","type":"string"},{"const":"done","type":"string"}]},"priority":{"anyOf":[{"const":"low","type":"string"},{"const":"medium","type":"string"},{"const":"high","type":"string"}]},"projectId":{"examples":[1],"type":"number"}}}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAllTasksResponse"}}}}},"security":[{"bearer":[]}],"tags":["/tasks"]}},"/v1/tasks/{id}":{"get":{"operationId":"getTaskByIdV1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetTaskByIdResponse"}}}}},"security":[{"bearer":[]}],"tags":["/tasks"]},"put":{"operationId":"updateTaskV1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskBody"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskResponse"}}}}},"security":[{"bearer":[]}],"tags":["/tasks"]},"delete":{"operationId":"deleteTaskV1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteTaskResponse"}}}}},"security":[{"bearer":[]}],"tags":["/tasks"]}}},"info":{"title":"Nila API","description":"Core API for Nila platform","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterBody":{"examples":[{"email":"john.doe@example.com","password":"securePassword123"}],"type":"object","properties":{"email":{"format":"email","examples":["john.doe@example.com"],"type":"string"},"password":{"minLength":6,"description":"Password must be at least 6 characters long","examples":["securePassword123"],"type":"string"}},"required":["email","password"]},"RegisterResponse":{"type":"object","properties":{"user":{"type":"object","properties":{"id":{"examples":[1],"type":"number"},"email":{"format":"email","examples":["john.doe@example.com"],"type":"string"}},"required":["id","email"]},"token":{"description":"JWT token for authentication","examples":["eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."],"type":"string"}},"required":["user","token"]},"LoginBody":{"examples":[{"email":"john.doe@example.com","password":"securePassword123"}],"type":"object","properties":{"email":{"format":"email","examples":["john.doe@example.com"],"type":"string"},"password":{"minLength":6,"description":"Password must be at least 6 characters long","examples":["securePassword123"],"type":"string"}},"required":["email","password"]},"LoginResponse":{"type":"object","properties":{"user":{"type":"object","properties":{"id":{"examples":[1],"type":"number"},"email":{"format":"email","examples":["john.doe@example.com"],"type":"string"}},"required":["id","email"]},"token":{"description":"JWT token for authentication","examples":["eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."],"type":"string"}},"required":["user","token"]},"CreateProjectBody":{"examples":[{"name":"Task Management System","description":"A system to manage tasks and track their progress"}],"type":"object","properties":{"name":{"minLength":3,"maxLength":255,"examples":["Task Management System"],"type":"string"},"description":{"maxLength":1000,"examples":["A system to manage tasks and track their progress"],"type":"string"}},"required":["name"]},"CreateProjectResponse":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","name","description"]},"GetAllProjectsResponse":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","name","description"]}},"GetProjectByIdResponse":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","name","description"]},"UpdateProjectBody":{"examples":[{"name":"Updated: Task Management System","description":"Updated: A system to manage tasks and track their progress"}],"type":"object","properties":{"name":{"minLength":3,"maxLength":255,"examples":["Updated: Task Management System"],"type":"string"},"description":{"maxLength":1000,"examples":["Updated: A system to manage tasks and track their progress"],"type":"string"}}},"UpdateProjectResponse":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","name","description"]},"DeleteProjectResponse":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","name","description"]},"CreateTaskBody":{"examples":[{"title":"Implement user authentication","description":"Implement JWT based authentication with refresh token support","status":"todo","priority":"high","projectId":1}],"type":"object","properties":{"title":{"minLength":1,"maxLength":255,"examples":["Implement user authentication"],"type":"string"},"description":{"examples":["Implement JWT based authentication with refresh token support"],"type":"string"},"status":{"anyOf":[{"const":"todo","type":"string"},{"const":"in_progress","type":"string"},{"const":"done","type":"string"}]},"priority":{"anyOf":[{"const":"low","type":"string"},{"const":"medium","type":"string"},{"const":"high","type":"string"}]},"projectId":{"examples":[1],"type":"number"}},"required":["title","projectId"]},"CreateTaskResponse":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]},"status":{"anyOf":[{"const":"todo","type":"string"},{"const":"in_progress","type":"string"},{"const":"done","type":"string"}]},"priority":{"anyOf":[{"const":"low","type":"string"},{"const":"medium","type":"string"},{"const":"high","type":"string"}]},"projectId":{"type":"number"}},"required":["id","title","description","status","priority","projectId"]},"GetAllTasksResponse":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]},"status":{"anyOf":[{"const":"todo","type":"string"},{"const":"in_progress","type":"string"},{"const":"done","type":"string"}]},"priority":{"anyOf":[{"const":"low","type":"string"},{"const":"medium","type":"string"},{"const":"high","type":"string"}]},"projectId":{"type":"number"}},"required":["id","title","description","status","priority","projectId"]}},"GetTaskByIdResponse":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]},"status":{"anyOf":[{"const":"todo","type":"string"},{"const":"in_progress","type":"string"},{"const":"done","type":"string"}]},"priority":{"anyOf":[{"const":"low","type":"string"},{"const":"medium","type":"string"},{"const":"high","type":"string"}]},"projectId":{"type":"number"}},"required":["id","title","description","status","priority","projectId"]},"UpdateTaskBody":{"type":"object","properties":{"title":{"minLength":1,"maxLength":255,"examples":["Updated: Implement user authentication"],"type":"string"},"description":{"examples":["Updated: Implement JWT based authentication with refresh token support"],"type":"string"},"status":{"anyOf":[{"const":"todo","type":"string"},{"const":"in_progress","type":"string"},{"const":"done","type":"string"}]},"priority":{"anyOf":[{"const":"low","type":"string"},{"const":"medium","type":"string"},{"const":"high","type":"string"}]}},"examples":[{"title":"Updated: Implement user authentication","description":"Updated: Implement JWT based authentication with refresh token support","status":"in_progress","priority":"high"}]},"UpdateTaskResponse":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]},"status":{"anyOf":[{"const":"todo","type":"string"},{"const":"in_progress","type":"string"},{"const":"done","type":"string"}]},"priority":{"anyOf":[{"const":"low","type":"string"},{"const":"medium","type":"string"},{"const":"high","type":"string"}]},"projectId":{"type":"number"}},"required":["id","title","description","status","priority","projectId"]},"DeleteTaskResponse":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]},"status":{"anyOf":[{"const":"todo","type":"string"},{"const":"in_progress","type":"string"},{"const":"done","type":"string"}]},"priority":{"anyOf":[{"const":"low","type":"string"},{"const":"medium","type":"string"},{"const":"high","type":"string"}]},"projectId":{"type":"number"}},"required":["id","title","description","status","priority","projectId"]}}}}